chore: always display task option above steps (#2835)
