+++ ui/user/src/lib/components/tasks/OnDemand.svelte
@@ -2,7 +2,7 @@
 	import { autoHeight } from '$lib/actions/textarea';
 	import { tooltip } from '$lib/actions/tooltip.svelte';
 	import type { OnDemand } from '$lib/services';
-	import { Minus, Plus } from 'lucide-svelte';
+	import { Plus, Trash2 } from 'lucide-svelte';
 
 	interface Props {
 		onDemand?: OnDemand;
@@ -11,6 +11,8 @@
 
 	let { onDemand = $bindable(), readOnly }: Props = $props();
 	let order = $state<string[]>([]);
+
+	let hasParams = $derived(!!Object.keys(onDemand?.params ?? {}).length);
 
 	$effect(() => {
 		for (const key in onDemand?.params ?? {}) {
@@ -24,12 +26,17 @@
 <div class="flex flex-col gap-4">
 	<h4 class="text-base font-medium">Arguments</h4>
 	<p class="text-gray text-sm">
-		Reference these values in your steps using <span class="font-mono text-black dark:text-white"
-			>$VAR</span
-		> syntax
+		If your task has arguments, you’ll be prompted to enter them when running the task manually from
+		the UI and if the task is invoked from a chat thread or another task, the agent will supply the
+		arguments based on its current context and instructions.
+	</p>
+	<p class="text-gray text-sm">
+		You can reference these arguments in your steps using
+		<span class="font-mono text-black dark:text-white">$name</span> syntax, like “Crawl $website and
+		find pages it links to.”
 	</p>
 
-	{#if onDemand?.params}
+	{#if hasParams}
 		<table class="w-full text-left">
 			<thead class="text-sm">
 				<tr>
@@ -60,7 +67,7 @@
 						<td class="w-1/2 pr-4 align-bottom">
 							<textarea
 								use:autoHeight
-								bind:value={onDemand.params[order[i]]}
+								bind:value={onDemand!.params![order[i]]}
 								class="text-input w-full resize-none py-2.5 align-bottom"
 								disabled={readOnly}
 								placeholder="Add a good description"
@@ -73,14 +80,16 @@
 									class="icon-button"
 									onclick={() => {
 										const key = order[i];
-										order = order.filter((k) => k !== key);
+
 										if (onDemand?.params) {
 											delete onDemand.params[key];
 										}
+
+										order = order.filter((k) => k !== key);
 									}}
 									use:tooltip={'Remove Argument'}
 								>
-									<Minus class="size-5" />
+									<Trash2 class="size-5" />
 								</button>
 							</td>
 						{/if}
@@ -89,6 +98,7 @@
 			</tbody>
 		</table>
 	{/if}
+
 	{#if !readOnly}
 		<div class="self-end">
 			<button
